Refactor error handling and logging in Oracle module
- Cleaned up error handling in `errors.rs` by removing unnecessary whitespace.
- Improved logging messages in `fetchers/mod.rs` and `fetchers/stock.rs` for better clarity.
- Reorganized module imports in `lib.rs` and `models.rs` for consistency.
- Enhanced readability by formatting code in `oracle.rs` and `main.rs`.
- Streamlined API calls and error handling in stock price fetching methods.
- Updated price fetching logic to ensure better error reporting and handling.
