+++ CODE_REFACTORING.md
@@ -1,0 +1,153 @@
+# Code Refactoring - Kanari API
+
+## Overview
+
+The original `api.rs` file (805 lines) has been successfully refactored into a modular structure for better maintainability and organization.
+
+## New Structure
+
+```
+crates/kanari-api/src/
+├── api.rs              # Main router and server initialization (76 lines)
+├── auth.rs             # Authentication functions
+├── database.rs         # Database operations and initialization
+├── models.rs           # Data structures and API models
+├── handlers/
+│   ├── mod.rs         # Handler module exports
+│   ├── health.rs      # Health check endpoints
+│   ├── price.rs       # Price-related endpoints
+│   └── user.rs        # User management endpoints
+└── lib.rs             # Module exports
+```
+
+## Modules Description
+
+### 1. `api.rs` (Main Router)
+
+- Contains the main `create_router()` function
+- Server initialization `start_api_server_with_shared_oracle()`
+- Application state definition (`AppState`)
+- **Size**: Reduced from 805 lines to 76 lines
+
+### 2. `auth.rs` (Authentication)
+
+- Token validation (`validate_token()`)
+- Token creation (`create_monthly_token()`)
+- Authentication utilities
+
+### 3. `database.rs` (Database Operations)
+
+- Database initialization (`initialize_database()`)
+- Database pool creation (`create_db_pool()`)
+- Database table setup
+
+### 4. `models.rs` (Data Models)
+
+- API response structures (`ApiResponse<T>`)
+- Request/Response models for all endpoints
+- Serializable/Deserializable structs
+
+### 5. `handlers/` (Endpoint Handlers)
+
+#### `handlers/health.rs`
+
+- `health_check()` - Health monitoring endpoint
+
+#### `handlers/price.rs`
+
+- `get_price()` - Individual asset price
+- `get_all_prices()` - All prices for asset type
+- `list_symbols()` - Available trading symbols
+- `get_stats()` - Oracle statistics
+- `update_prices()` - Force price updates
+
+#### `handlers/user.rs`
+
+- `register_user()` - User registration
+- `login_user()` - User authentication
+- `list_users()` - Admin user listing
+- `get_user_profile()` - User profile retrieval
+- `delete_user_account()` - Account deletion
+
+## Benefits of Refactoring
+
+### 1. **Modularity**
+
+- Each module has a single responsibility
+- Easy to locate and modify specific functionality
+- Better code organization
+
+### 2. **Maintainability**
+
+- Smaller, focused files are easier to understand
+- Reduced cognitive load when working on specific features
+- Better separation of concerns
+
+### 3. **Testing**
+
+- Individual modules can be unit tested independently
+- Mock dependencies more easily
+- Better test coverage
+
+### 4. **Reusability**
+
+- Authentication logic can be reused across handlers
+- Database functions are centralized
+- Models can be shared between modules
+
+### 5. **Collaboration**
+
+- Multiple developers can work on different modules simultaneously
+- Reduced merge conflicts
+- Clear ownership of functionality
+
+## API Endpoints (Unchanged)
+
+The refactoring maintains all existing API functionality:
+
+``` 
+GET  /health                     - Health check
+GET  /price/{type}/{symbol}      - Get specific price
+GET  /prices/{type}              - Get all prices for type
+GET  /symbols                    - List available symbols
+GET  /stats                      - Oracle statistics
+POST /update/{type}              - Force update prices
+POST /users/register             - Register new user
+POST /users/login                - User login
+GET  /users/list                 - List all users
+GET  /users/profile              - Get user profile
+POST /users/delete               - Delete user account
+```
+
+## Usage
+
+The refactored code maintains the same public API interface:
+
+```rust
+use kanari_api::api::start_api_server_with_shared_oracle;
+
+// Usage remains the same
+start_api_server_with_shared_oracle(shared_oracle, 3000).await?;
+```
+
+## Migration Notes
+
+- **Import changes**: Update any direct imports from `api.rs` to use the new module structure
+- **No API changes**: All HTTP endpoints remain the same
+- **Database schema**: No changes to database tables or queries
+- **Configuration**: Environment variables and config remain unchanged
+
+## File Sizes Comparison
+
+| Original | Refactored |
+|----------|------------|
+| `api.rs`: 805 lines | `api.rs`: 76 lines |
+| | `auth.rs`: 41 lines |
+| | `database.rs`: 47 lines |
+| | `models.rs`: 87 lines |
+| | `handlers/health.rs`: 17 lines |
+| | `handlers/price.rs`: 195 lines |
+| | `handlers/user.rs`: 338 lines |
+| | `handlers/mod.rs`: 6 lines |
+
+**Total**: Same functionality, better organized across multiple focused modules.
