+++ crates/kanari-api/src/handlers/mod.rs
@@ -1,0 +1,7 @@
+pub mod health;
+pub mod price;
+pub mod user;
+
+pub use health::*;
+pub use price::*;
+pub use user::*;
