+++ crates/kanari-api/src/lib.rs
@@ -1,1 +1,5 @@
 pub mod api;
+pub mod auth;
+pub mod database;
+pub mod handlers;
+pub mod models;
