+++ crates/kanari-oracle/src/lib.rs
@@ -1,5 +1,5 @@
-pub mod oracle;
-pub mod fetchers;
-pub mod models;
 pub mod config;
 pub mod errors;
+pub mod fetchers;
+pub mod models;
+pub mod oracle;
